Datastructures #1 - PoC of standardizing a couple of fields
This commits changes the type of two fields in envelope.h from struct List to
to struct STailQHead. This new structure represents a singly-linked tail queue
implemented using FreeBSD's queue.h (imported).

This is a proof of concept to see how we like using macros to implement basic
datastructures (singly and doubly linked lists and tail queues) and algorithms
(traversal, removal, insertion, ...).

I encourage everybody to have a look at the result and comment. Is this the
direction we would like to take?

Issue #374